Update CUDA custom op unit tests to account for recent ORT change (#6971)
